Fix loop counter data type to support more shift registers
- Changed loop counter type from uint8_t to uint16_t to handle more than 255 bits.
- Updated shiftRegister function to support up to 8 daisy chained shift registers.
- Improved scalability by using uint16_t for the loop counter to avoid overflow issues with a high number of registers.

+++ ECUAL/SHIFT_REGISTER/SHIFT_REGISTER.c
@@ -10,10 +10,10 @@
 
 #include "SHIFT_REGISTER.h"
 
-void shiftRegister(uint8_t data){
+void shiftRegister(uint64_t data){
 	const shiftReg_CfgType* config = &shiftReg_CfgParam;
 	uint8_t numBits = config->numRegister * 8;
-	for(uint8_t i = 0; i < numBits; i++){
+	for(uint16_t i = 0; i < numBits; i++){
 		HAL_GPIO_WritePin(config->DATA_GPIO_Port, config->DATA_Pin, (data >> i) & 0x01);
         HAL_GPIO_WritePin(config->CLOCK_GPIO_Port, config->CLOCK_Pin, GPIO_PIN_SET);
         HAL_GPIO_WritePin(config->CLOCK_GPIO_Port, config->CLOCK_Pin, GPIO_PIN_RESET);
